added recent projects container
